+++ src/main/java/com/historycode/ui/page/streetCodePage/components/AdditionalInfoComponent.java
@@ -1,6 +1,7 @@
-package com.historycode.ui.page.streetCodePage;
+package com.historycode.ui.page.streetCodePage.components;
 
 import com.historycode.ui.component.BaseComponent;
+import com.historycode.ui.page.streetCodePage.modals.AdditionalInfoModal;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
