+++ src/main/java/com/historycode/ui/page/streetCodePage/components/carousels/FactsCarousel.java
@@ -1,14 +1,15 @@
-package com.historycode.ui.page.streetCodePage;
+package com.historycode.ui.page.streetCodePage.components.carousels;
 
+import com.historycode.ui.page.streetCodePage.components.FactCardComponent;
+import com.historycode.ui.page.streetCodePage.components.PaginationComponent;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class FactsCarousel extends BaseCarousel{
+public class FactsCarousel extends BaseCarousel {
     @FindBy(xpath = ".//div[@class='interestingFactSlide']")
     private List<WebElement> factCardNodes;
