+++ src/main/java/com/historycode/ui/page/streetCodePage/components/carousels/PartnersCarousel.java
@@ -1,14 +1,14 @@
-package com.historycode.ui.page.streetCodePage;
+package com.historycode.ui.page.streetCodePage.components.carousels;
 
+import com.historycode.ui.page.streetCodePage.components.PartnerLogoComponent;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class PartnersCarousel extends BaseCarousel{
+public class PartnersCarousel extends BaseCarousel {
     @FindBy(xpath = ".//div[@class='partnerItem']")
     private List<WebElement> partnerLogoNodes;
